+++ app/javascript/packs/react.js
@@ -8,6 +8,6 @@
 // layout file, like app/views/layouts/application.html.erb
 
 // Support component names relative to this directory:
-var componentRequireContext = require.context("components", true)
+var componentRequireContext = require.context("react_components", true)
 var ReactRailsUJS = require("react_ujs")
 ReactRailsUJS.useContext(componentRequireContext)
