+++ app/javascript/react_components/Main.js
